pyproject.toml
[project]
name = "ThermanPrinter"
version = "0.0.1"
description = "Enables printing simple web content (news, puzzles, etc.) to any ESC/POS thermal printer."
authors = [
    {name = "Oliver Nemec", email = "28827159+onemec@users.noreply.github.com"},
]

dependencies = [
    "dominate>=2.9.1",
    "libusb>=1.0.29",
    "pillow>=12.0.0",
    "playwright>=1.56.0",
    "py-sudoku>=2.0.0",
    "pydantic>=2.12.5",
    "python-escpos[all]>=3.1",
    "pyusb>=1.3.1",
    "requests>=2.32.5",
    "typer>=0.20.0",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "GPL-3.0-only"}

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
